feat: handling empty fields
Also Known As - If this field is empty there's no need to show this on the screen. Visibility is set to Gone.

Origin - If origin is not populated, Show text - "It's a mystery!"

Image - If image load runs into issues, display a default image that shows "No Image Available"
This happened with Schwarama while developing